+++ test/accountsResolvers.test.js
@@ -1,31 +1,5 @@
-import { graphqlTestCall } from "./graphqlTestCall";
-import mongoose from 'mongoose';
-import { MongoMemoryServer } from 'mongodb-memory-server';
+import { graphqlTestCall } from "./helper/graphqlTestCall";
 import { User } from "../models/users";
-
-let mongoServer;
-const opts = { }; 
-
-beforeAll(async () => {
-  mongoServer = new MongoMemoryServer();
-  const mongoUri = await mongoServer.getUri();
-  await mongoose.connect(mongoUri, opts, (err) => {
-    if (err) console.error(err);
-  });
-});
-
-beforeEach(async () => {
-  const collections = mongoose.connection.collections;
-  for (const key in collections) {
-    const collection = collections[key];
-    await collection.deleteMany({});
-  }
-});
-
-afterAll(async () => {
-  await mongoose.disconnect();
-  await mongoServer.stop();
-});
 
 const createUserMutation = `
   mutation createUser($firstName: String!, $lastName: String!, $email: String!, $password: String!) {
@@ -165,6 +139,22 @@ describe("queries", () => {
     const expectedUser = { firstName: "Mihran", lastName: "Guyumjyan", email: "mihran@gmail.com", userId: userResponse.data.getUser.userId };
   
     expect(userResponse.data.getUser).toEqual(expectedUser);
+  });
+
+  it("should not get user information without userId", async () => {
+    const testUser = { firstName: "Mihran", lastName: "Guyumjyan", email: "mihran@gmail.com", password: "pass" };
+      
+    await graphqlTestCall(createUserMutation, {
+      email: testUser.email,
+      password: testUser.password,
+      lastName: testUser.lastName,
+      firstName: testUser.firstName
+    });
+
+    const userResponse = await graphqlTestCall(getUserQuery, {} );
+
+    expect(Array.isArray(userResponse.errors)).toBeTruthy();
+    expect(userResponse.errors).not.toHaveLength(0);
   });
 
   it("should add age to users collection", async () => {
