+++ test/helper/graphqlTestCall.js
@@ -1,9 +1,7 @@
-import "babel-polyfill";
 import { graphql } from "graphql";
 import { makeExecutableSchema } from "graphql-tools";
-
-import { typeDefs } from "../resolvers/typeDefs";
-import { resolvers } from "../resolvers/index";
+import { typeDefs } from "../../resolvers/typeDefs";
+import { resolvers } from "../../resolvers/index";
 
 const schema = makeExecutableSchema({ typeDefs, resolvers });
